refactor: restructure project and expose Go toolchain by default
Closes #40

Move go-scrape utility to `scrape/` directory and update flake outputs
so that packages.default and apps.default expose the Go toolchain
instead of the internal scrape tool.

BREAKING CHANGE: `nix build` and `nix shell` now provide Go toolchain
instead of go-scrape. Use .#go-scrape to access the scrape utility.
