+++ .github/workflows/auto-update.yml
@@ -25,7 +25,7 @@ jobs:
 
       - name: Check GoDev Website
         run: |
-          echo "LATEST_GO_VERSION=$(nix run . -- detect)" >> $GITHUB_ENV
+          echo "LATEST_GO_VERSION=$(nix run .#go-scrape -- detect)" >> $GITHUB_ENV
 
       - name: Check Manifest Exists
         run: |
@@ -36,7 +36,7 @@ jobs:
       - name: Generate Manifest
         if: ${{ env.MANIFEST_EXISTS != 'true' }}
         run: |
-          nix run . -- generate ${{ env.LATEST_GO_VERSION }} --output manifests
+          nix run .#go-scrape -- generate ${{ env.LATEST_GO_VERSION }} --output manifests
 
       - name: Import GPG key
         if: ${{ env.MANIFEST_EXISTS != 'true' }}
